# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_ReportPreview_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReportPreview"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ReportPreview]",
        "type": "object"
      },
...
      "Envelope_ToolStats_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToolStats"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ToolStats]",
        "type": "object"
      },
...
      "Envelope_list_ToolStats__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ToolStats"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[list[ToolStats]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportPreview": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "sample": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sample",
            "type": "array"
          }
        },
        "required": [
          "row_count",
          "columns",
          "sample"
        ],
        "title": "ReportPreview",
        "type": "object"
      },
...
      "ToolStats": {
        "properties": {
          "call_count": {
            "title": "Call Count",
            "type": "integer"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "mcp_server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mcp Server"
          },
          "projects": {
            "default": 0,
            "title": "Projects",
            "type": "integer"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "unique_sessions": {
            "title": "Unique Sessions",
            "type": "integer"
          }
        },
        "required": [
          "tool_name",
          "call_count",
          "unique_sessions"
        ],
        "title": "ToolStats",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only REST API over the local Token Telemetry database. Binds to 127.0.0.1; see docs/API.md for conventions (envelopes, filters, pagination, time zones).",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "summary": "Local-first token, tool, skill and MCP observability for Claude Code.",
    "title": "Token Telemetry API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reports/export": {
      "get": {
        "operationId": "export_report_api_v1_reports_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "requests",
              "pattern": "^(requests|projects)$",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "pattern": "^(csv|json|ndjson)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "description": "Optional maximum number of rows",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional maximum number of rows",
              "title": "Limit"
            }
          },
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download a report as CSV or JSON (streamed, no row cap)",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/preview": {
      "get": {
        "operationId": "preview_report_api_v1_reports_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "requests",
              "pattern": "^(requests|projects)$",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReportPreview_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Row count + first rows of a report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/tools": {
      "get": {
        "operationId": "list_tools_api_v1_tools_get",
        "parameters": [
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_ToolStats__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tool call counts",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/{tool_name}": {
      "get": {
        "operationId": "tool_detail_api_v1_tools__tool_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "tool_name",
            "required": true,
            "schema": {
              "title": "Tool Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ToolStats_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One tool's stats",
        "tags": [
          "tools"
        ]
      }
    },
...
  }
}